+++ lib/jwt/jwa.rb
@@ -40,6 +40,11 @@ module JWT
 
         algorithm
       end
+
+      def resolve_and_sort(algorithms:, preferred_algorithm:)
+        algs = Array(algorithms).map { |alg| JWA.resolve(alg) }
+        algs.partition { |alg| alg.valid_alg?(preferred_algorithm) }.flatten
+      end
     end
   end
 end